Fixed table and graphs.
Using flat trackings for table and aggregated for graphs.
