+++ src/api/delay.js
@@ -1,1 +1,1 @@
-export default 100;
+export default 2000;
