+++ src/components/dashboard/DashboardPage.js
@@ -6,12 +6,16 @@ import * as dashboardActions from "../../actions/dashboardActions";
 import DashboardTable from "./DashboardTable";
 import StackedBarChart from "./StackedBarChart";
 import SimpleLineChart from "./SimpleLineChart";
+import _ from "lodash";
+import { getFormattedDate } from "../../utils/dateHelper";
 
 export class DashboardPage extends React.Component {
 
     render() {
       const { aggregates } = this.props;
       const { assets } = this.props;
+
+      let groupedAggregates = GroupTrackings(aggregates);
       return (
         <div>
           <h1>Dashboard</h1>
@@ -21,14 +25,51 @@ export class DashboardPage extends React.Component {
                 <td colSpan="2"><DashboardTable rows={aggregates} assets={assets} /></td>
             </tr>
             <tr>
-              <td><StackedBarChart data={aggregates} /></td>
-              <td><SimpleLineChart data={aggregates} /></td>
+              <td><StackedBarChart data={groupedAggregates} /></td>
+              <td><SimpleLineChart data={groupedAggregates} /></td>
             </tr>
             </tbody>
           </table>
         </div>
       );
     }
+}
+
+function GroupTrackings(allTrackings) {
+  var trackingsByTime = _.groupBy(allTrackings, "trackingTime");
+  //console.log(JSON.stringify(trackingsByTime));
+
+  let final = [];
+
+  for (var time in trackingsByTime) {
+    let totalCostPrice = 0;
+    let totalMarketPrice = 0;
+    let totalYieldValue = 0;
+    trackingsByTime[time].forEach(tracking => {
+      totalCostPrice = totalCostPrice + tracking.costPrice;
+      totalMarketPrice = totalMarketPrice + tracking.marketPrice;
+      totalYieldValue = totalYieldValue + tracking.yieldValue;
+    });
+    final.push({
+      "trackingTime": time,
+      "trackingTimeShort": getFormattedDate(new Date(time)),
+      "costPrice": totalCostPrice,
+      "marketPrice": totalMarketPrice,
+      "yieldValue": totalYieldValue,
+      "yieldPercent": (totalYieldValue / totalCostPrice) * 100,
+      "children": trackingsByTime[time]
+    });
+  }
+
+  final.sort(function(a, b){
+    var x = a.trackingTimeShort.toLowerCase();
+    var y = b.trackingTimeShort.toLowerCase();
+    if (x < y) {return -1;}
+    if (x > y) {return 1;}
+    return 0;
+  });
+  //console.log(JSON.stringify(final));
+  return final;
 }
 
     DashboardPage.propTypes = {
